requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.3.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
google-auth-oauthlib==1.2.0
google-auth==2.27.0
google-generativeai==0.3.2
pillow==10.2.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-magic==0.4.27
beanie==1.29.0 